requirements.txt
streamlit
groq
PyPDF2
python-docx
python-dotenv
fpdf